feat: add delete shipment request and response interfaces; create reusable dialog component

+++ app/api/shipment/[trackingId]/route.ts
@@ -1,12 +1,16 @@
 import { NextResponse } from "next/server";
 import { connectToDatabase } from "@/lib/mongodb";
 import Shipment from "@/models/shipment";
-import { handleApiError, handleResourceNotFoundError } from "@/lib/utils";
+import {
+  handleApiError,
+  handleAuthError,
+  handleResourceNotFoundError,
+} from "@/lib/utils";
 import { ShipmentResponse } from "@/types/shipment";
 
 export async function GET(
   req: Request,
-  { params }: { params: Promise<{ trackingId: string }> },
+  { params }: { params: Promise<{ trackingId: string }> }
 ) {
   try {
     await connectToDatabase();
@@ -15,6 +19,7 @@ export async function GET(
 
     const shipment = await Shipment.findOne({ trackingId });
     if (!shipment) {
+      console.log("Shipment not found", trackingId);
       return handleResourceNotFoundError("Shipment not found");
     }
 
@@ -30,10 +35,23 @@ export async function GET(
 
 export async function DELETE(
   req: Request,
-  { params }: { params: Promise<{ trackingId: string }> },
+  { params }: { params: Promise<{ trackingId: string }> }
 ) {
   try {
     await connectToDatabase();
+
+    // Authenticate request
+    const authHeader = req.headers.get("authorization");
+    const token = authHeader?.replace("Bearer ", "");
+
+    if (!token) {
+      return handleAuthError("Auth token missing");
+    }
+
+    if (token !== process.env.AUTH_TOKEN) {
+      return handleAuthError("Invalid auth token");
+    }
+
     const { trackingId } = await params;
 
     const deleted = await Shipment.findOneAndDelete({ trackingId });
@@ -43,7 +61,7 @@ export async function DELETE(
 
     return NextResponse.json(
       { status: "success", data: { message: "Shipment deleted successfully" } },
-      { status: 200 },
+      { status: 200 }
     );
   } catch (error) {
     return handleApiError(error);
